[Feat, Fix]: 카테고리별 총 포스트 수 연결, 기타 버그 수정
- 카테고리별 총 포스트 수를 백엔드로부터 받아온다.
  - 이후 카테고리 페이지에서 카테고리명(총포스트 수) 형식으로 표현.

- 카테고리 페이지에서 변경사항 저장 후 "바로 새로고침하면" 변경사항이 저장이 안되는 문제
  문제 발생 이유는 mutation loading/mutation success/mutation 성공 후 refetch하는 과정에서 새로고침을 하게 되면 서버측 DB에는 저장이 되었지만, 클라이언트측 퍼시스트 로컬스토리지에는 저장이 안되게 된다.
  즉 데이터의 일관성이 깨지게 된다.
  해결 방법은 useEffect에 beforeunload이벤트를 추가하고, 의존성에 mutation loading, refetching를 추가.
  mutation loading 및 refetching 중에 새로고침하면  localStorage.removeItem("REACT_QUERY_OFFLINE_CACHE");를 통해 캐시를 삭제한다.
  이후에 다시 카테고리 페이지에 접근하게 되면, 캐시 데이터가 남아있지 않게 되어 새로운 데이터를 불러와 캐시 데이터를 업데이트 한다.
  단 여기서 의존성에 mutation success를 포함하게 되면, 기존에 설계했던 방식과 맞지 않아 mutation success는 제외. mutation success를 제외해도 원하는 대로 동작.

- 글 삭제, 글 생성, 글 수정 완료시에 똑같이 캐시 데이터를 삭제하고 이후에 카테고리 페이지에 접근했을 때 새로운 신선한 데이터를 불러옴으로써 DB에 있는 최근 데이터와 무결성을 유지할 수 있음.

- 2단 카테고리에서 새 카테고리를 추가할 때 자식 카테고리는 중복 검사 로직이 적용 안되는 현상 수정 완료.

- Category.tsx에서 중복되는 코드 함수로 처리해 가독성 및 유지보수를 높임.

- 카테고리 삭제 시 포스트에 하나라도 연결되어 있는 카테고리라면 삭제 불가능 UI 처리.	백엔드에서 응답으로 건네 준 postCount 활용.

- 카테고리 삭제 시 해당 카테고리에 하나 이상의 자식 카테고리가 있다면 삭제 불가능 처리 로직 수정 -> 기존 코드에서 백엔드로부터 응답받은 childCount 추가 활용.
- 글 수정/ 삭제 페이지에서 카테고리 목록을 불러올 때 React quill와 충돌 가능성을 고려해서 로직을 적용.
  - 첫 번째 시도 방법: 맨 초기에 불러오는 카테고리 데이터는 React quill에디터와 충돌 가능성이 거의 없다고 봐도 무방하다 생각하여 초기에 불러오는것만 State를 적용
  이후에 수정할때는 react quill내부 에디터 상태와 충돌 가능성이 있기 때문에 ref를 사용. -> 이 방법은 새로운 카테고리로 선택하는 UI가 작동하지 않음
  - 두 번째 시도한 해결 방법은 Ref는 DOM을 직접 제어해야 하기 때문에 html select 및 option태그를 동적으로 생성한 후 초기에 불러온 카테고리 데이터를 할당.
  이후에 UI적으로 카테고리를 수정할때 처음 데이터를 불러오는것과 같이 ref를 사용.
  최종적으로 문제 해결.
  - 추가로 글 수정시 성공하면 router.replace를 사용했는데 이 경우 router replace로 설정한 경로로 이동후에, 현재 컴포넌트가 재실행되는 문제가 발생함.
  - 이에 따라 글 수정시 성공하면 로컬스토리지에 저장된 캐시 데이터를 삭제했는데, 다시 캐시 데이터가 생기는 의도치 않은 동작 발생
  - 해결 방법은 window.location 사용하여 해당 경로로 이동 후에, 현재 컴포넌트 재실행을 막음으로써 해결
  - 이유를 계속 생각해봤는데, 아마도 router는 클라이언트측 네비게이션이라서 이것과 연관이 있을거라 생각 됨. 왜냐하면 수정 페이지 관련된 컴포넌트를 완전한 비제어 컴포넌트로 전환했기 때문에 현재 컴포넌트가 재실행될 이유가 없다고 생각.
  - 반면 window.location 방식은 완전한 새로고침 방식이기 때문에 해결할 수 있었다고 생각.
- 현재 회원가입, 로그인 페이지 세부 구현 진행 중.

+++ next-blog/src/customHooks/useDeletePost.ts
@@ -1,12 +1,12 @@
 import { refreshToken } from "@/utils/refreshToken";
-import { useRouter } from "next/navigation";
+import { useEffect, useState } from "react";
 import { useMutation, useQueryClient } from "react-query";
 
 import "react-toastify/dist/ReactToastify.css";
 
-const deletePost: (postId: string, accessToken: string | boolean, userIdentifier: string) => Promise<Response> = async (
+const deletePost: (postId: string, accessToken: string | null, userIdentifier: string) => Promise<Response> = async (
     postId: string,
-    accessToken: string | boolean,
+    accessToken: string | null,
     userIdentifier: string
 ): Promise<Response> => {
     return fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}${process.env.NEXT_PUBLIC_BACKEND_PATH}/${userIdentifier}/posts/${postId}`, {
@@ -18,12 +18,20 @@ const deletePost: (postId: string, accessToken: string | boolean, userIdentifier
     });
 };
 
-// 내부함수인 useMutation를 실행할때 postId, userIdentifier, accessToken은 클로져로 참조 가능. 
+// 내부함수인 useMutation를 실행할때 postId, userIdentifier, accessToken은 클로져로 참조 가능.
 function useDeletePost(postId: string, userIdentifier: string) {
-    const accessToken = localStorage.getItem("access_token") ?? false;
+    const [accessToken, setAccessToken] = useState<string | null>(null);
+
+    useEffect(() => {
+        const token = localStorage.getItem("access_token");
+        setAccessToken(token);
+    }, []);
 
     return useMutation(
         async () => {
+
+            if (accessToken === null) return;
+
             let response = await deletePost(postId, accessToken, userIdentifier);
 
             if (!response.ok) {
